Updated git submodules
Project: genius master b1510188a72471c592e49328417cf6393dddafc2

Fix failing LockManagerTest (NoClassDefFoundError)

The LockManagerTest is currently failing on clean master with:
java.lang.NoClassDefFoundError:
org/opendaylight/controller/md/sal/binding/impl/BindingToNormalizedNodeCodec

The cause of this problem is that genius.testutils has a dependency to
sal-binding-broker-impl's attached <type>test-jar, only - but not the
main artifact (where BindingToNormalizedNodeCodec is).

It's easy to fix like this - but I don't really understand, and don't
have time to further investigate, how this could have suddenly broken...
presumably someone changed a dependency in some POM in controller or
somewhere which had this as a side effect.  Be that as it may - it works
again now with this change.

Change-Id: Ic46d0f78cb89b0a704f1e603f8dd115f46987cb1
Signed-off-by: Michael Vorburger <vorburger@redhat.com>