Updated git submodules
Project: netvirt stable/carbon 0c6ecdf4707cbccd82c3ae1b34dff18d1724404e

BUG9094 Bind to last SF interface if origin node

If the last SF is in one of the possible origin nodes for the
classified traffic, bind the classifier on it's port. This is to be
to handle the possible scenario where origin, last SF and destination
of traffic are on the same node. On this scenario, SFC on last hop will
not de-encapsulate the traffic, instead it will be handed off to the
dispatcher table for the egress classifier function to handle it. To be
able to do this, it needs to be bound to the last SF port.

Change-Id: I97f60e8cc409ea42c830e408b4df91051b50689c
Signed-off-by: Jaime Caamaño Ruiz <jaime.caamano.ruiz@ericsson.com>
(cherry picked from commit 9509d783b6c9fe5a348e0f39e86b8ae72cc53c3f)