Updated git submodules
Project: netconf stable/nitrogen 1e8350d9c5aa9558f0c128625165d4532a20409a

Transition ListenerAdapter to ClusteredDOMDataTreeListener

Solves this: https://bugs.opendaylight.org/show_bug.cgi?id=9147
This is in general an issue with ListenerAdapter which does not
currently work in a clustered deployment.

ListenerAdapterTest needed to be adjusted for two reasons:
1. There was a bug in ListenerAdapter where if leafNodesOnly
was set the leafNodesOnly logic was not applied in delete
notifications. The new ListenerAdapter code fixes that so
ListenerAdapterTest needed to be adjusted.
2. In DomDataTreeEtc, merge operations cause the key node
to count as "WRITE"-en for notifications even if it was not
changed. ListenerAdapterTest was adjusted accordingly. Bug
for this here: https://bugs.opendaylight.org/show_bug.cgi?id=9179

Change-Id: I3c40e02551d8f04354489e5ffb8c66f111bb3b96
Signed-off-by: Josh <jhershbe@redhat.com>


Add unit tests for ListenerAdapter

Test that json generated by ListenerAdapter is correct
for regular notifications and "leaf nodes only" notifications.

Change-Id: I62fa22b1e30ca6a8632b3d2614558e9ced0abf54
Signed-off-by: Josh <jhershbe@redhat.com>