Updated git submodules
Project: sfc master 11b92f30d4c57fa8edfb66eab09333eae0e47eed

vxlan_tool was checking only two bytes of the mac

When opening a raw sockets, the forwarded packets are also consumed. Therefore
we need a way to drop the generated packets so that they are not processed
again. The way to do this was by compating the interface's mac address and the
destination mac address of the packet, however, the comparison only checked
two bytes and thus there were situations in which the tool misbehaved. This
patch compares all the bytes of the mac

Change-Id: I151b7cc5d379e38f86d6583c6aad41ebe2e62011
Signed-off-by: Manuel Buil <mbuil@suse.com>