Capture Jira issue ID from commit message
Update release notes generator script to handle migration from
Bugzilla to Jira. Presently, the script only adds a link to bug id.

Parse and add an issue id if available in the commit message. If the
issue id is not available retrive the issue id from Jira using REST
API's.

Change-Id: Ifb82b9fd7909d9f4a59123b441088bff0a4be547
Signed-off-by: Anil Belur <abelur@linuxfoundation.org>