Capture JIRA issue ID from commit message
Update release notes generator script to handle migration from
Bugzilla to JIRA. Presently, the script only adds a link to bug id.

Parse and add an issue id if available in the commit message. If the
issue id is not available then retrive the issue id from JIRA using
REST API's.

Also improve the format of the console output to make it more readalbe.

JIRA: RELENG-593
Change-Id: Ib18c92b2b346d327f008791332cdd45bcee754ef
Signed-off-by: Anil Belur <abelur@linuxfoundation.org>