Updated git submodules
Project: genius master 05a5ad26f149cce43a419e381346ee5ae7ca7cf5

LockManager unlock() with RetryingManagedNewTransactionRunner

lock() and tryLock() should ideally also be switched to
(Retrying)ManagedNewTransactionRunner and FutureRpcResults, but that's a
little trickier, because of the return true/false in readWriteLock().

This is a partial counter proposal to
https://git.opendaylight.org/gerrit/#/c/61526/ related to
https://jira.opendaylight.org/browse/GENIUS-86.  It's not a complete
counter proposal; the changes other than the retry on unlock should
still be considered.

Change-Id: Ie6588a0314828a3d1e88c741b7f8d511afb29e12
Signed-off-by: Michael Vorburger <vorburger@redhat.com>