Update patch set 5
Patch Set 5: Code-Review-1

(1 comment)

As mentioned in the previous comment, not sure why but I'm getting `curl: (7) Couldn't connect to server` when I try to run this script. Would be good to print what it's trying to connect to so we can check for issues.

Also the curl loop runs VERY slow. We should figure out how we can code this to run more quickly.

Patch-set: 5
Label: Code-Review=-1