Update patch set 6
Patch Set 6:

> Are you running this locally or through a job?

I'm running locally. It's happening in between the loop so looks like there likely commits creating a URL that's incorrect but we cannot confirm this without more debug output from the script. Printing the curl download status is not useful so we should probably hide that and instead print out the URLs it's trying to query.

Patch-set: 6