Updated git submodules
Project: netvirt master 3963ca8e4fcf9328559bf7e6d79f84b905eb21a9

Remove static VrfEntryListener.isOpenStackVniSemanticsEnforced

VrfEntryListener defines a static isOpenStackVniSemanticsEnforced
which is set in the ctor via the IElanService. FindBugs flags this
as a violation although it technically is safe b/c VrfEntryListener
is a singleton. However statics in this manner aren't good practice.
Users of the static flag should instead inject and obtain the flag
from IElanService.

One user was the static FibUtil class. FibUtil was converted to a
singleton. It also had static methods that take the DataBroker
and IdManagerService. Thee services are now injected and the methods
changed to non-static to make it cleaner. All users of these methods
now inject the FibUtil instance.

Change-Id: Ifecd9511568e263545ca7ce1c486483f67f9fe85
Signed-off-by: Tom Pantelis <tompantelis@gmail.com>


Fix FindBugs violations and enable enforcement in fibmanager-impl

See in-line comments

Change-Id: I7eeaa287693922b26d92bef840680858b5519845
Signed-off-by: Tom Pantelis <tompantelis@gmail.com>