Updated git submodules
Project: aaa master e1b9ab15e3b3dc0d598395432411e8d4f3fc2ff6

Collapse aaa feature repositories

AAA advertises two feature repositories.  Historically, this was done based
on the fact that it was hoped to pull out a ton of the stuff surrounding
non-shiro implementation in downstream distributions by just discluding the
non-shiro feature repository.  However, it appears that this has limited use,
and primarily serves as a distraction.  Furthermore, there are more
intelligent ways to perform such an operation.

The first part of this change renames the "authn" module to "aaa", because
the contained features repository is "features-aaa" and is not specific
to authentication.

The second part moves odl-aaa-shiro to the aaa module.  This is done in
preparation to remove features-aaa-shiro, which will no longer be offered.

features-aaa-shiro still contains the odl-aaa-shiro feature for now.
After all downstreams are migrated to utilize features-aaa instead, it
will be completely removed as a module.  This is a multistep process and
should not affect downstream consumers.

Change-Id: Ie084b622b1e58c661fee910e85a204436bf23f3e
Signed-off-by: Ryan Goulding <ryandgoulding@gmail.com>