Updated git submodules
Project: aaa master 47d13b19b94c1542d0fbd36a205d3a974a202efd

Merge "Rid AAA source of binary images"

Rid AAA source of binary images

From long ago we referenced some images in our README.md and directly
stored them in source.  Thats a bad idea for a number of reasons.  This
simply gets rid of the references to the one remaining image and the
source/output graphics that were stored in source.  If we want them back,
we can figure out a way to upload them to somewhere more appropriate.

The one existing accurate diagram, direct_authn.png, was really simple
to understand anyway.  If we want, we can restore it somewhere new,
put it in ascii-art, or just describe it better.  Also, it is probably
better to consolidate that sort of information in the docs project.

Change-Id: I104c32eb844cf4031f26f6108aeb8b98d13b9401
Signed-off-by: Ryan Goulding <ryandgoulding@gmail.com>