Update patch set 1
Patch Set 1:

Not sure if this is an improvement over the solution proposed by I8ae38e2ce65ddab1bfc56e6244b9a19ae494703e but wanted to try an idea that was floating in my head. Feel free to abandon or take elements of this implementation if determined that it is an inferior solution.

Patch-set: 1
Label: Verified=0