Update patch set 2
Patch Set 2:

> (2 comments)

Its good practice to lock reads/writes to file when you are accessing files parallelly. Using flock does not decrease the performance of the script in any way, and ensures consistency. I am unable to reproduce the issue at this point and can remove it.  However, I can remove this from the issue is not reproducible ...

Patch-set: 2
Label: Verified=0