Update patch set 3
Patch Set 3:

> Its good practice to lock reads/writes to file when you are accessing files parallelly.

Depends if it's actually necessary or not. "echo" is guarenteed to be an atomic write as long as the length of the data is shorter than PIPE_BUF / BUFSIZ. Our simple cache db is very short writes that does not even come close to the atomic limit imo. Calling flock in this case is an unnecessary step.

Some details here to explain this https://stackoverflow.com/a/9927415

Patch-set: 3