Update patch set 1
Patch Set 1:

I talked with An about this. We think it should be moved to RelEng and based on the JJB job start I have there. He has some very cool ideas actually about this class as a lamda that runs on fancy-cool infra that I totally didn't understand, but for maintainability reasons maybe better to keep with the JJB/ODL infra pattern for now.

An, I said we should just split this and not abandon it, but I actually didn't notice this is in autorelease not builder, so do actually need to abandon it and manually move.

Patch-set: 1