Update patch set 1
Patch Set 1:

As mentioned in the reverted patch:

This is wrong. This is a heading for projects with NO noteworthy changes. See:

http://docs.opendaylight.org/en/stable-nitrogen/release-notes/release-notes-nitrogen-sr1.html

We are now stating that all projects did not contribute any commits to a release have noteworthy changes.

It does produce an empty list if every project had changes in a release though. Perhaps the real fix is to check if an empty list is produced and print a "None" statement.

Patch-set: 1
Reviewer: Gerrit User 2759 <2759@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0