Update patch set 3
Patch Set 3: Code-Review+1

I'm fine to merge this, as we can iterate on it over
time as people have cycles and needs.

I would have liked to see some docs in the functions/header
to understand what I was seeing. I was digging through the
code to understand the output.

maybe some stdout to go with the final json to explain what
we are looking at?

anyway, all just good-to-have things.

This was VERY cool to have and use. Big props to An for giving
us this.

Patch-set: 3
Label: Code-Review=+1