Update patch set 3
Patch Set 3:

> (3 comments)
 > 
 > We should run flake8 against this file just to make sure it passes.

it won't pass, pep8 etc. I think we were just trying to get something in that worked. This file is a good start but it needs work to be refactored and cleaned.

I will take a quick pass on the pep8 stuff and push another revision.

Patch-set: 3
Label: Verified=0