Update patch set 1
Patch Set 1:

just question - is it not part of the release at this point? can't recall what the work flow is to move projects to MRI. Though guess it doesn't matter. as mdsal was included as an mri artifact in the bumps right?

Patch-set: 1
Reviewer: Gerrit User 773 <773@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=0