Remove Bugzilla code from generator
This code was originally used to support the ODL Bugzilla
to JIRA transition. This code is no longer needed as we do
not expect any current projects to be referencing Bugzilla
IDs in their commit messages any longer.

This also resolves issue with BGPCEP-# JIRAs being picked
up as PCEP-# incorrectly. This is likely a result of the
Bugzilla regex being bugged.

Issue: INFRA-146
Signed-off-by: Thanh Ha <zxiiro@gmail.com>
Change-Id: I81589d0a4825e2c9615648c9e117f62b51dd6d56